experiment with pytorch Dataset class
much slower than my original implementation. Need to learn whether it is useful when not running in parallelized setups. Seems like a lot of extra work for the processors unless the data is so big that it is important to only load and use pieces in memory at a time (like the tutorial on image datasets).